remove webpack.config.js and move parts to storybook main.js
